Create a libbz2.so.1 symlink
The soname of Bzip2 is libbz2.so.1.0, so this is what tools should use
when linking against bz2. However, some tools incorrectly use
"libbz2.so.1". By providing this symlink, we can make them work.

(Apparently, using a soname with major.minor is somewhat unusual and the
normal way would indeed have been to use libbz2.so.1.)

Debian does this as well in their bzip2 package, see
https://salsa.debian.org/debian/bzip2/blob/0c9335755bcd9519bfc36cf57eb987603b013f46/debian/rules#L73